progressWknd
Added some complication to the last part of the point down algorithm to conver things to derivates and go from there. Doesnt seem to work, things just go to the start of the cue. should do serial printout to see whats happening.
Also added some RTC / NTP time setting stuff. Set up interrupts, but haven't tested that the GPIO updates. And then added the update leds skeleton but with a functional blacking out of sections based on the time.